+++ IMU.h
@@ -25,12 +25,12 @@ public:
     // array[0] = system status
     // array[1] = system test results
     // array[2] = system error
-    void getSystemStatus(uint8_t* array);
+    void getSystemStatus(uint16_t* array);
     //Populates int array where:
     // array[0] = X heading
     // array[1] = Y heading
     // array[2] = Z heading
-    void getXYZ(uint8_t* array);
+    void getXYZ(uint16_t* array);
 
 private:
     Adafruit_BNO055 bno;
